deduplicate and merge askxdd.py and xdd.py
